Enable hwloc support
This commit enables hwloc support by addind build dependency to
libhwloc-dev. Unfortunately, upstream autotools macro that checks
for hwloc headers also adds rpath in LDFLAGS and this has the effect
to make lintian very angry. That is why a new patch has been added to
fix the macro and avoid usage of rpath when --with-rpath flag is not
set. This patch, for being fruitful requires autoconf to be run after
being applied. It is the reason why autoreconf has been enable in
d/rules.